1,调整BUG日志手机策略; 2,商家销售数据展示;

+++ app/src/main/java/com/lrcall/utils/StringTools.java
@@ -145,7 +145,7 @@ public class StringTools
 	 * @param pricce 以分为单位的价格
 	 * @return
 	 */
-	public static String getPrice(int pricce)
+	public static String getPrice(long pricce)
 	{
 		return String.format("%.2f", (double) ((double) pricce / 100));
 	}
